pyproject.toml
[project]
name = "retail-chain"
version = "0.1.0"
description = "Веб приложение с API-интерфейсом и admin-панелью."
authors = [
    {name = "Vik Vyaz",email = "nightronin@yandex.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-coverage (>=0.0,<0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["retail_chain", "users"]
norecursedirs = ["management"]